Move the progress validation logic to its own class
